Expose Pathtracker, Endgamer constructors. Solve type inference problems
The type inference in the previous approach was very fragile. The
problem was probably some internal Julia limits with kwargs and type
inference. (Named tuples should solve that). To resolve that I had to
pass through kwargs with the splice syntax and add a filter function.
That's the same approach as previously (and the reason that the problem
didn't occur in the last version). In the same swoop I also made the
construction of Pathtracker and Endgamer more accessible.