Change the safety condition that a path is going to infinity
Previously we just used the absolute value of the homogenous coordinate.
But this can yield wrong results if one coordinate of the solution is
much larger than all others. A less sensitive check is to require
a certain minimal norm of the affine "solution" before declaring it at
infinity. This is now `minimal_maxnorm=1e3` (compard to `maxnorm=1e5`).