Change convergence criteria in Newton iteration
Previously we used the residual as the convergence criteria.
This is has the problem that it is scaling dependent and also not really
interesting for the quality of the solution. For (x-2)^5 1.8 has only
an accuracy of 0.2 but the residual would be 3.2e-9. Thus we switch
to the size of the update in the Newton step, which assuming we are
in the radius of quadratic convergence (which is reasonable),
gives us a good approximation of the number of correct digits.

As an consequence of this change, we now have more paths that fail,
which is expected since we where overly optimistic previously.