Clean up 'cond' measurement:
Rename `cond` in PathTracker and `newton_update_error` in newton.jl to 
`lost_digits` and track the log10 value of the value previous denoted by 
`cond`. This should reflect the true meaning better.

This also removes the iterative refinement routines as it turns out that 
the we cannot really make use of iterative refinement since either the 
accuracy is sufficient (2 or 3 digits accuracy is enough) or we are too 
ill-conditioned and iterative refinement doesn't converge anymore.

This also removes DoubleFloats for now again.