Improve precision of internal step size in Tracker
So far, the step size logic in Tracker (using ComplexLineSegment) 
assumed to track from length(segment) to 0. This had the advantage of 
being able to make tiny steps close to 0. However, for polyhedral 
homotopy we also need to make tiny steps at the beginning, i.e., we 
would need  to track from 0 to length(segment). This is now accomplished 
dinamically